LayoutTests/fast/loader/cache-encoding-expected.txt

First load a script with a wrong charset then again with the right one. Second attempt should work and 'scriptSuccess' should be set to true. 'successfullyParsed' will be undefined.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS scriptSuccess is true
PASS scriptSuccess is true
PASS successfullyParsed is true

TEST COMPLETE
